Add configuration for MQTT base topic and update node IDs; introduce time synchronization in publisher

+++ load_sampling/include/Common.h
@@ -17,7 +17,7 @@ struct SignalSnapshot
 
 struct PowerChannelSnapshot : SignalSnapshot
 {
-    float    sum_p;   // sum of v*i
+    double   sum_p;   // sum of v*i, still in ADC units, but can be used for power factor or real power if voltage is sampled.
 };
 struct NodeStatsSnapshot
 {
